Move slugify to identifier utils
