+++ f/common_logic/data_conversion.py
@@ -1,11 +1,8 @@
 import csv
 import json
 import logging
-import re
-import unicodedata
 import xml.etree.ElementTree as ET
 from pathlib import Path
-from typing import Any
 
 import filetype
 import fiona
@@ -673,32 +670,3 @@ def smart_xml_to_geojson(path: Path):
     from f.connectors.smart.smart_patrols import parse_smart_patrol_xml
 
     return parse_smart_patrol_xml(path)
-
-
-def slugify(value: Any, allow_unicode: bool = False) -> str:
-    """A safe slugify utility.
-
-    - Converts to str, normalizes unicode, optionally keeps unicode.
-    - Returns an ASCII-ish slug of the input suitable for file names.
-    - Returns 'unnamed' for empty inputs.
-
-    Source: https://github.com/django/django/blob/main/django/utils/text.py#L453
-    """
-    value = "" if value is None else str(value)
-    if not value:
-        return "unnamed"
-
-    if allow_unicode:
-        value = unicodedata.normalize("NFKC", value)
-    else:
-        value = (
-            unicodedata.normalize("NFKD", value)
-            .encode("ascii", "ignore")
-            .decode("ascii")
-        )
-
-    value = value.lower()
-    # keep alphanumerics, underscores, spaces and hyphens
-    value = re.sub(r"[^\w\s-]", "", value)
-    value = re.sub(r"[-\s]+", "-", value).strip("-_ ")
-    return value or "unnamed"
